Set up parse loop
This has to be done so that command will actually effect the environment.
Returning a list for
```
alias X=abc
X
```
would mean that we are unable to parse abc as the alias expansion
is done in the lexer. Returning them and having them affect a global state
object (Doesnt exist yet) will allow this to work.
This planned state object will also be used to store function defs
etc.